BoundsObj

An object of results from computing bounds

BoundsObj(lbs, ubs, ub.quantiles, lb.quantiles, ub.qte, lb.qte, att = NULL,
  probs)

Arguments

lbs
A vector of the lower bounds for each value in the support of the outcome

ubs
A vector of the upper bounds for each value in the support of the outcome

ub.quantiles
A vector of the same length as probs that contains the upper bound of the quantiles of the counterfactual distribution of untreated potential outcomes for the treated group

lb.quantiles
A vector of the same length as probs that contains the lower bound of the quantiles of the counterfactual distribution of untreated potential outcomes for the treated group

ub.qte
The point estimate of the upper bound for the QTE

lb.qte
The point estimate of the lower bound for the QTE

att
The ATT is point identified under the assumptions required by the bounds method

probs
A vector of values between 0 and 1 to compute the QTET at