bounds

bounds estimates bounds for the Quantile Treatment Effect on the Treated (QTET) using the method of Fan and Yu (2012).

bounds(formla, xformla = NULL, t, tmin1, tname, data, idname,
  probs = seq(0.05, 0.95, 0.05))

Arguments

formla
The formula y ~ d where y is the outcome and d is the treatment indicator (d should be binary), d should be equal to one in all time periods for individuals that are eventually treated

xformla
A optional one sided formula for additional covariates that will be adjusted for. E.g ~ age + education. Additional covariates can also be passed by name using the x paramater.

t
The 3rd time period in the sample. Treated individuals should be treated in this time period and untreated individuals should not be treated. The code attempts to enforce this condition, but it is good try to handle this outside the panel.qtet method.

tmin1
The 2nd time period in the sample. This should be a pre-treatment period for all individuals in the sample.

tname
The name of the column containing the time periods

data
A data.frame containing all the variables used

idname
The individual (cross-sectional unit) id name

probs
A vector of values between 0 and 1 to compute the QTET at

Value

A BoundsObj object

References

Fan, Yanqin and Zhengfei Yu. ``Partial Identification of Distributional and Quantile Treatment Effects in Difference-in-Differences Models.'' Economics Letters 115.3, pp.511-515, 2012.

Examples

## load the data
data(lalonde)

## Run the bounds method with no covariates
b1 <- bounds(re ~ treat, t=1978, tmin1=1975, data=lalonde.psid.panel,
  idname="id", tname="year")
summary(b1)

#> 
#> Bounds on the Quantile Treatment Effect on the Treated:
#> 		
#> tau	Lower Bound	Upper Bound
#>         tau	Lower Bound	Upper Bound
#>        0.05	     -51.72	          0
#>         0.1	   -1220.84	          0
#>        0.15	    -1881.9	          0
#>         0.2	   -2601.32	          0
#>        0.25	   -2916.38	     485.23
#>         0.3	   -3080.16	     943.05
#>        0.35	   -3327.89	    1505.98
#>         0.4	   -3240.59	    2133.59
#>        0.45	   -2982.51	    2616.84
#>         0.5	   -3108.01	     2566.2
#>        0.55	   -3342.66	    2672.82
#>         0.6	    -3491.4	     3065.7
#>        0.65	   -3739.74	    3349.74
#>         0.7	   -4647.82	    2992.03
#>        0.75	   -4826.78	    3219.32
#>         0.8	    -5801.7	    2702.33
#>        0.85	   -6588.61	    2499.41
#>         0.9	   -8953.84	    2020.84
#>        0.95	  -14283.61	     397.04
#> 
#> Average Treatment Effect on the Treated:	2326.51