cic_attgt

cic_attgt

cic_attgt(gt_data, xformla = ~1, ...)

Arguments

gt_data

data that is "local" to a particular group-time average treatment effect

xformla

one-sided formula for covariates used in the propensity score and outcome regression models

...

extra function arguments; not used here

Value

pte::attgt_noif object. cic_attgt computes attgt using the CIC approach. It also returns distributions of observed outcomes for the treated group (F1), the counterfactual distribution of untreated potential potential outcomes for the treated group (F0), and the distribution of the treatment effect under the assumption of rank invariance over time (Fte) all through the extra_gt_returns argument to pte::attgt_noif object.