compute.panel.qte

compute.panel.qte uses third period of data, combined with Distributional Difference in Differences assumption (Fan and Yu, 2012) to point identify QTET.

compute.panel.qtet(formla, xformla = NULL, t, tmin1, tmin2, tname, x = NULL,
  data, dropalwaystreated = TRUE, idname, probs = seq(0.05, 0.95, 0.05),
  bootstrap.iter = FALSE, plot = FALSE, method = c("logit", "GMM",
  "semiparametric", "simulation", "direct"), ydiscrete = FALSE)

Arguments

formla

outcome variable on treatment

t

last time period

tmin1

middle time period

tmin2

initial time period

x

additional covariates if using propensity score reweighting technique

dropalwaystreated

boolean indicating whether in true panel data context (when treatment can occur in any period) whether or not previously treated observations should be dropped from the sample. This functionality is not currently implemented

idname

an id that identifies individual units over time

probs

the values at which the quantile treatment effects should be computed

bootstrap.iter

boolean passed that is passed in when this method is used to compute standard errors

plot

Binary indicating whether or not to plot the object

method

How to estimate the propensity score

ydiscrete

Used to indicate whether the outcome has any discrete parts in the distribution

Value

QTE object