ci.qtet

The ci.qtet method implements estimates the Quantile Treatment Effect on the Treated (QTET) under a Conditional Independence Assumption (sometimes this is called Selection on Observables) developed in Firpo (2007). This method using propensity score re-weighting and minimizes a check function to compute the QTET. Standard errors (if requested) are computed using the bootstrap.

ci.qtet(formla, xformla = NULL, x = NULL, weights = NULL, data,
  probs = seq(0.05, 0.95, 0.05), se = TRUE, iters = 100, alp = 0.05,
  plot = FALSE, method = "logit", retEachIter = FALSE, seedvec = NULL,
  indsample = TRUE, printIter = FALSE, pl = FALSE, cores = 2)

Arguments

formla

The formula y ~ d where y is the outcome and d is the treatment indicator (d should be binary)

xformla

A optional one sided formula for additional covariates that will be adjusted for. E.g ~ age + education. Additional covariates can also be passed by name using the x paramater.

x

Vector of covariates. Default is no covariates

weights

an additional vector of sampling weights

data

The name of the data.frame that contains the data

probs

A vector of values between 0 and 1 to compute the QTET at

se

Boolean whether or not to compute standard errors

iters

The number of iterations to compute bootstrap standard errors. This is only used if se=TRUE

alp

The significance level used for constructing bootstrap confidence intervals

plot

Boolean whether or not the estimated QTET should be plotted

method

Method to compute propensity score. Default is logit; other option is probit.

retEachIter

Boolean whether or not to return list of results from each iteration of the bootstrap procedure

seedvec

Optional value to set random seed; can possibly be used in conjunction with bootstrapping standard errors.

indsample

Binary variable for whether to treat the samples as independent or dependent. This affects bootstrap standard errors. In the job training example, the samples are independent because they are two samples collected independently and then merged. If the data is from the same source, usually should set this option to be FALSE.

printIter

For debugging only; should leave at default FALSE unless you want to see a lot of output

pl

boolean for whether or not to compute bootstrap error in parallel. Note that computing standard errors in parallel is a new feature and may not work at all on Windows.

cores

the number of cores to use if bootstrap standard errors are computed in parallel

Value

QTE object

References

Firpo, Sergio. ``Efficient Semiparametric Estimation of Quantile Treatment Effects.'' Econometrica 75.1, pp. 259-276, 2015.

Examples

## Load the data
data(lalonde)

##Estimate the QTET of participating in the job training program;
##This is the no covariate case.  Note: Because individuals that participate
## in the job training program are likely to be much different than
## individuals that do not (e.g. less experience and less education), this
## method is likely to perform poorly at estimating the true QTET
q1 <- ci.qtet(re78 ~ treat, x=NULL, data=lalonde.psid, se=FALSE,
 probs=seq(0.05,0.95,0.05))
summary(q1)

#> 
#> Quantile Treatment Effect:
#> 		
#> tau	QTE
#> 0.05	     0.00
#> 0.1	     0.00
#> 0.15	 -4388.53
#> 0.2	 -8783.29
#> 0.25	-11171.92
#> 0.3	-12434.94
#> 0.35	-13780.90
#> 0.4	-15401.86
#> 0.45	-15827.89
#> 0.5	-16472.18
#> 0.55	-17375.87
#> 0.6	-18023.28
#> 0.65	-18448.70
#> 0.7	-19188.63
#> 0.75	-19878.24
#> 0.8	-20816.84
#> 0.85	-22761.15
#> 0.9	-23843.88
#> 0.95	-27184.77
#> 
#> Average Treatment Effect:	-15204.78

##This estimation controls for all the available background characteristics.
q2 <- ci.qtet(re78 ~ treat,
 xformla=~age + I(age^2) + education + black + hispanic + married + nodegree,
 data=lalonde.psid, se=FALSE, probs=seq(0.05, 0.95, 0.05))

#> Error in `[.data.frame`(data, , x): undefined columns selected

summary(q2)

#> Error in summary(q2): object 'q2' not found